Authentication with JWT, decorators and guards

+++ src/restaurants/restaurants.controller.ts
@@ -8,7 +8,9 @@ import {
   Delete,
   Query,
   UseInterceptors,
-  UploadedFiles
+  UploadedFiles,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { Query as ExpressQuery } from 'express-serve-static-core';
@@ -16,49 +18,73 @@ import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { CurrentUser } from '../auth/decorators/current-user.decorator';
+import { RolesGuard } from '../auth/guards/roles.guard';
+import { Roles } from '../auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
   constructor(private restaurantsService: RestaurantsService) {}
 
   @Get()
-  async getAllRestaurants(@Query() query: ExpressQuery): Promise<Restaurant[]>{
+  async getAllRestaurants(@Query() query: ExpressQuery): Promise<Restaurant[]> {
     return this.restaurantsService.findAll(query);
   }
 
   @Post()
+  @UseGuards(AuthGuard(), RolesGuard)
+  @Roles('admin', 'user')
   async createRestaurant(
     @Body()
-    restaurant: CreateRestaurantDto
+    restaurant: CreateRestaurantDto,
+    @CurrentUser() user,
   ): Promise<Restaurant> {
-    return this.restaurantsService.create(restaurant);
+    return this.restaurantsService.create(restaurant, user);
   }
 
   @Get(':id')
   async getRestaurantById(
     @Param('id')
-    id: string
+    id: string,
   ): Promise<Restaurant> {
     return this.restaurantsService.findById(id);
   }
 
   @Put(':id')
+  @UseGuards(AuthGuard())
   async updateRestaurant(
     @Param('id')
     id: string,
     @Body()
-    restaurant: UpdateRestaurantDto
+    restaurant: UpdateRestaurantDto,
+    @CurrentUser() user,
   ): Promise<Restaurant> {
-    await this.restaurantsService.findById(id);
+    const res = await this.restaurantsService.findById(id);
+
+    if (res.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException(
+        'You are not allowed to update this restaurant',
+      );
+    }
+
     return this.restaurantsService.updateById(id, restaurant);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard())
   async deleteRestaurant(
     @Param('id')
-    id: string
+    id: string,
+    @CurrentUser() user,
   ): Promise<{ deleted: boolean }> {
     const restaurant = await this.restaurantsService.findById(id);
+
+    if (restaurant.user.toString() !== user._id.toString()) {
+      throw new ForbiddenException(
+        'You are not allowed to delete this restaurant',
+      );
+    }
     const deleteImages = await this.restaurantsService.deleteImages(
       restaurant.images,
     );
@@ -71,11 +97,12 @@ export class RestaurantsController {
   }
 
   @Put('upload/:id')
+  @UseGuards(AuthGuard())
   @UseInterceptors(FilesInterceptor('files'))
   async uploadFiles(
     @Param('id')
     id: string,
-    @UploadedFiles() files: Array<Express.Multer.File>
+    @UploadedFiles() files: Array<Express.Multer.File>,
   ) {
     await this.restaurantsService.findById(id);
     const res = await this.restaurantsService.uploadImages(id, files);
